chore: split project command helpers
